Matrix transform API

+++ silo-core/src/models/group.rs
@@ -1,6 +1,9 @@
+use serde::Serialize;
+
 /// Groups multiple subjects together.
-#[derive(Debug)]
+#[derive(Debug, Serialize)]
+#[serde(rename_all = "camelCase")]
 pub struct Group {
     /// The group's unique ID.
-    pub id: i16,
+    pub id: i32,
 }
