+++ silo-db/src/actor.rs
@@ -24,11 +24,11 @@ impl DbActor {
 
 /// Inserts a Subject into the DB.
 #[derive(Message)]
-#[rtype(result = "Result<i16, DatabaseError>")]
+#[rtype(result = "Result<i32, DatabaseError>")]
 pub struct InsertSubject(&'static models::Subject);
 
 impl Handler<InsertSubject> for DbActor {
-    type Result = ResponseFuture<Result<i16, DatabaseError>>;
+    type Result = ResponseFuture<Result<i32, DatabaseError>>;
 
     fn handle(&mut self, msg: InsertSubject, ctx: &mut Self::Context) -> Self::Result {
         let query = self.service.insert_subject(msg.0);
@@ -39,11 +39,11 @@ impl Handler<InsertSubject> for DbActor {
 
 /// Inserts a SubjectTrait into the DB.
 #[derive(Message)]
-#[rtype(result = "Result<i16, DatabaseError>")]
+#[rtype(result = "Result<i32, DatabaseError>")]
 pub struct InsertSubjectTrait(&'static models::SubjectTrait);
 
 impl Handler<InsertSubjectTrait> for DbActor {
-    type Result = ResponseFuture<Result<i16, DatabaseError>>;
+    type Result = ResponseFuture<Result<i32, DatabaseError>>;
 
     fn handle(&mut self, msg: InsertSubjectTrait, ctx: &mut Self::Context) -> Self::Result {
         let query = self.service.insert_subject_trait(msg.0);
@@ -54,11 +54,11 @@ impl Handler<InsertSubjectTrait> for DbActor {
 
 /// Inserts a Group into the DB.
 #[derive(Message)]
-#[rtype(result = "Result<i16, DatabaseError>")]
+#[rtype(result = "Result<i32, DatabaseError>")]
 pub struct InsertGroup(&'static models::Group);
 
 impl Handler<InsertGroup> for DbActor {
-    type Result = ResponseFuture<Result<i16, DatabaseError>>;
+    type Result = ResponseFuture<Result<i32, DatabaseError>>;
 
     fn handle(&mut self, msg: InsertGroup, ctx: &mut Self::Context) -> Self::Result {
         let query = self.service.insert_group(msg.0);
